requirements.txt
# Lightweight dependencies for Vercel deployment
# Only includes dependencies for: greeting, todays-date, get_weather, google_sheets tools

fastapi>=0.115.0
uvicorn>=0.30.0
optimizely-opal-opal-tools-sdk>=0.1.0
pydantic>=2.0.0
httpx>=0.27.0
